Initial reorginization
