--- Old Stuff/README.md
@@ -1,2 +1,0 @@
-# SailBot
-Main codebase for SailBot 2019-20
