--- controller/LoRaReceiver/messages.pb.c
@@ -1,36 +1,0 @@
-/* Automatically generated nanopb constant definitions */
-/* Generated by nanopb-0.4.0-dev */
-
-#include "messages.pb.h"
-
-/* @@protoc_insertion_point(includes) */
-#if PB_PROTO_HEADER_VERSION != 40
-#error Regenerate this file with the current version of nanopb generator.
-#endif
-
-PB_BIND(SkipperCommand, SkipperCommand, AUTO)
-
-
-PB_BIND(SailCommand, SailCommand, AUTO)
-
-
-PB_BIND(RudderCommand, RudderCommand, AUTO)
-
-
-PB_BIND(Mode, Mode, AUTO)
-
-
-PB_BIND(BaseToBoat, BaseToBoat, AUTO)
-
-
-PB_BIND(SailStatus, SailStatus, AUTO)
-
-
-PB_BIND(PowerStatus, PowerStatus, AUTO)
-
-
-PB_BIND(BoatToBase, BoatToBase, AUTO)
-
-
-
-/* @@protoc_insertion_point(eof) */
