+++ deprecated/constants.py
