+++ deprecated/image_analysis_test.py
