+++ deprecated/windvane.py
