+++ experimental/image_analysis/image_analysis_gui.py
