+++ sailbot/onboard/boat_hack.py
