+++ sailbot/onboard/constants.py
