+++ sailbot/onboard/windvane.py
