+++ sailbot/proto/README.md
@@ -1,0 +1,1 @@
+Protobuf definitions for wireless comms
